[internal] allow some attributes to be type cast before save
Why?
- this will allow us to control when we type cast a column and when we
dont. this will be useful with rails 5 in this scenario: a column is of
the type decimal and we are trying to save its model to the database,
but that column is `nil`. if we read that column before type casting it we
would get `""`, but after type casting it would be `nil`. `nil` is what
we want because we can save that to the db, but we can't save an empty
string to the database in a decimal column with the stricter sql mode.
- this will allow us to control how we save specific columns with this
gem until we move to rails 6 which has its own bulk insert mechanism.

What?
- we now can pass in an array of column names for the models trying to
be saved. the columns that we pass in will be type cast before we try to
save.